[Feature #163969676] Refactor v1 structure and tests
- Moved validator to upper folder
- Refactored unclear status codes
