+++ .env
@@ -7,5 +7,5 @@ export FLASK_ENV='development'
 
 #database uris as per config
 
-export DEVELOPMENT_DATABASE_URI=''
+export DEVELOPMENT_DATABASE_URI='postgresql://postgres:politico_dev@localhost:5432/politico_db_development'
 export PRODUCTION_DATABASE_URI=''
