+++ api/v1/models/petition_model.py
@@ -1,6 +1,6 @@
 from . import Model
 from time import localtime
-from api.v1.validator import PetitionValidator
+from api.validator import PetitionValidator
 
 petitions = []
