+++ tests/v1tests/test_petition_view.py
@@ -38,7 +38,7 @@ class PetitionEndpointsTestCase(BaseTestCase):
         Tests Petition Creation Unsuccessful
         """
         response = self.client.post('api/v1/petitions', data=json.dumps(self.petition_invalid_key))
-        self.assertEqual(response.status_code, 403)
+        self.assertEqual(response.status_code, 400)
         self.assertEqual(response.json['error'], "Invalid Request ,Missing Data")
 
     def test_petition_invalidated(self):
